remove array dataloader altogether all logic in loader base
